pyproject.toml
[tool.poetry]
name = "django-ml-api"
version = "0.1.0"
description = ""
authors = ["sunilregmi233 <sunilregmi223@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.6.0"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "39.0.0"
defusedxml = "0.7.1"
dj-rest-auth = "2.2.6"
django = "4.1.5"
django-allauth = "0.52.0"
django-cors-headers = "3.13.0"
djangorestframework = "3.14.0"
drf-yasg = "1.21.4"
idna = "3.4"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.2"
markupsafe = "2.1.1"
oauthlib = "3.2.2"
packaging = "23.0"
pillow = "9.4.0"
pycparser = "2.21"
pyjwt = "2.6.0"
python3-openid = "3.2.0"
pytz = "2022.7"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
ruamel-yaml = "0.17.21"
ruamel-yaml-clib = "0.2.7"
six = "1.16.0"
sqlparse = "0.4.3"
uritemplate = "4.1.1"
urllib3 = "1.26.14"
rest-framework-simplejwt = "^0.0.2"
rest-social-auth = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
